fixup??? grep: follow conventions for printing paths w/ unusual chars
It is easy to be fooled by the Bash included in Git for Windows, which
leads you to believe that quotes are valid parts of file names.

On Windows, they are not. But Cygwin (which is the base of MSYS2, which
is the POSIX emulation layer used by that Bash) only _pretends_ that it
is a valid file name character. In reality, it will map the character
into the private Unicode page. Cygwin knows about this. The rest of
Windows applications (including Git for Windows), however, does not.

As a consequence, `>\"with\ quotes\"` will claim to succeed, but the
file on disk will have Unicode characters in place of those quotes that
literally no application but Cygwin ones can handle, and this leads to
those beautiful new tests to fail.

Let's just use the prereq we introduced to guard precisely against this
problem: `FUNNYNAMES`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>